각 페이지마다 클래스 파일 작성
